cambios

+++ src/controllers/movieControllers.js
@@ -4,7 +4,9 @@ import { Genre } from "../models/Genre.js";
 export const getMovies = async (req, res) => {
   try {
     const movies = await Movie.findAll();
-    return res.status(200).json(movies);
+    movies.length > 0
+      ? res.status(200).json(movies)
+      : res.status(201).json("No hay películas creadas");
   } catch (error) {
     return res.status(500).json({ message: error.message });
   }
@@ -18,8 +20,8 @@ export const getMovieById = async (req, res) => {
         id,
       },
     });
-    if(!id){
-      return res.status(200).json("Movie does not exists")
+    if (!id) {
+      return res.status(200).json("Movie does not exists");
     }
     return res.status(200).json(movie);
   } catch (error) {
@@ -30,6 +32,13 @@ export const getMovieById = async (req, res) => {
 export const postMovie = async (req, res) => {
   try {
     const { image, title, date, rating, characterId } = req.body;
+    if (!image || !title || !date || !rating || !characterId) {
+      return res.status(200).json("Empty data is not allowed");
+    } else if (rating > 5) {
+      return res.status(404).json("Rating: Must be between 1 and 5");
+    } else if (!characterId) {
+      return res.status(404).json("Id inexistente");
+    }
     const newMovie = await Movie.create({
       image,
       title,
@@ -37,11 +46,6 @@ export const postMovie = async (req, res) => {
       rating,
       characterId,
     });
-    if(!image || !title || !date || !rating || !characterId){
-      return res.status(200).json("Empty data is not allowed")
-    } else if (rating > 5){
-      return res.status(200).json("Rating: Must be between 1 and 5")
-    }
     return res.status(200).json(newMovie);
   } catch (error) {
     return res.status(500).json({ message: error.message });
@@ -52,6 +56,8 @@ export const updateMovie = async (req, res) => {
   const { id } = req.params;
   const { image, title, date, rating } = req.body;
   try {
+    const movie = await Movie.findByPk(id);
+    console.log(movie);
     const movieUpdate = await Movie.update(
       { image, title, date, rating },
       {
@@ -60,7 +66,11 @@ export const updateMovie = async (req, res) => {
         },
       }
     );
-    return res.status(200).json(movieUpdate);
+    if (movieUpdate === movie) {
+      return res.status(404).json("No se han realizado cambios");
+    } else {
+      return res.status(200).json("Cambios realizados");
+    }
   } catch (error) {
     return res.status(500).json({ message: error.message });
   }
@@ -69,12 +79,16 @@ export const updateMovie = async (req, res) => {
 export const deleteMovie = async (req, res) => {
   const { id } = req.params;
   try {
-    const movieDelete = await Movie.destroy({
-      where: {
-        id,
-      },
-    });
-    return res.status(204).send(movieDelete);
+    if (id) {
+      const movieDelete = await Movie.destroy({
+        where: {
+          id,
+        },
+      });
+      return res.status(200).json("Movie deleted");
+    } else {
+      return res.status(404).json("Id not existing");
+    }
   } catch (error) {
     return res.status(500).json({ message: error.message });
   }
